+++ xrslam/src/xrslam/core/sliding_window_tracker.h
@@ -26,11 +26,27 @@ class SlidingWindowTracker {
 
     void refine_subwindow();
 
+    bool judge_track_status();
+    void fuse_imu_track();
+    bool check_frames_rpe(Track *track, const vector<3> &p);
+    void predict_RT(Frame *frame_i, Frame *frame_j, matrix<3> &R, vector<3> &t);
+    bool filter_parsac_2d2d(Frame *frame_i, Frame *frame_j,
+                            std::vector<char> &mask,
+                            std::vector<size_t> &pts_to_index);
+    void update_track_status();
+
+    std::vector<vector<3>> m_P3D;
+    std::vector<vector<2>> m_P2D;
+    std::vector<size_t> m_lens;
+    std::vector<int> m_indices_map;
+
     bool track();
 
     std::tuple<double, PoseState, MotionState> get_latest_state() const;
 
+    double m_th;
     std::unique_ptr<Map> map;
+    std::shared_ptr<Map> feature_tracking_map;
 
   private:
     std::shared_ptr<Config> config;